Moved MoveItem to it's own file Added X,Y to MoveItem
